chore(deps): bump eslint-plugin-disable-autofix for eslint v9 compat
Users have had issues with `eslint-plugin-disable-autofix` failing with @eslint v9. I have published an update to extend the compatibility to v9. It can still be used with v8 but I wanted to open this PR at least as a reminder for when the migration to v9 is completed since I noticed the config here is already done, if the plugin is not updated it will possibly throw an error!